Updated the Verify unit test so that it actually runs successfully in NUnit.
-->NOTE<--
The TestFixtureSetup involves starting up storescp.exe which is used as a DICOM server for the C-FIND test. The unit test assumes that storescp.exe exists in c:\windows. As well, the dcmnet.dll C++ library must be loadable by NUnit. In my local setup, I have put it in the NUnit execution directory.

Until I can figure how to better use the TestFixtureSetup in NUnit, this is the way it'll have to be.

The TestFixtureTearDown forcefully kills storescp.exe at this point.