R19. Added ImageCodec extension point.  New transfer syntaxes can now be supported by simply writing a new image codec plugin.
When a DicomImage is created, a static instance of an ImageCodecMap is created.  The map is populated with whatever implementations of IImageCodec are present.  The transfer syntax UID is used as the key to the map.  When pixel data is requested, the pixel data is decoded and returned.  A number of "identity" codecs, such as ImplicitVRLittleEndianCodec have been included for uncompressed pixel data.