Requirement: R787
Introduction of the Html preview, which uses the WebBrowser control, led to discovery that the Desktop framework was not actually calling Dispose() on workspace/shelf controls when they were removed from the parent control (this could be a DotNetMagic problem).  Added explicit calls to Dispose() when these controls are removed from the desktop.