Requirement: R787
Added Time property to the Platform class.  Platform.Time should be used instead of DateTime.Now to obtain an enterprise-wide value for the current time.  Internally, the property delegates to an extension which must implement an ITimeProvider interface.  If this extension is not found, the local machine time is used.