Requirement: R787
Refactored IScriptEngine so that it can create an IExecutableScript (an encapsulated script object that can be executed multiple times).  This was an attempt to improve the performance of the HTML view rendering.  Performance did not actually improve (because the Jscript engine has either already made the optimization internally, or because it cannot make the optimization), but this design is still a better design and so it has been retained.