Requirement: R89
Refactored DdlWriter to make use of extension points, allowing for additional ddl scripts to be inserted into the output.  The DdlWriter application will still write out the base scripts to define the table structure and populate enumerations as the first step.  Then it will process any extensions of DdlScriptGeneratorExtensionPoint to additional scripts.  This refactoring required moving a number of classes from ClearCanvas.Enterprise.Hibernate.DdlWriter to ClearCanvas.Enterprise.Hibernate.  The DdlWriter plugin is now just a single class, DdlWriterApplication, that is responsible for executing the script generation and writing it to a file.