Requirement: R883
Added the ability to specify, in ReadOperation and UpdateOperation attributes, a flag called PersistenceScopeOption which is either Requires or RequiresNew.  The default is Requires, meaning the operation requires a PersistenceContext and will inherit one if available, or create one if not.  RequiresNew will always create a new PersistenceContext regardless of whether one is already open.  RequiresNew is only needed in very specialized situations, where a service operation must execute in an isolated context (eg when generating an Accession number, we do not want to rollback the generator if the "place order" transaction fails, as this would allow the same Accession number to be obtained multiple times).