Requirement: R89
Added somewhat experimental BindableTreeView.  This class is analogous to TableView, in the sense that its purpose is to allow a tree view to be simply bound to a datasource exposed by an application component.  BindableTreeView can bind to a datasource of type ClearCanvas.Desktop.Trees.ITreeNode or ClearCanvas.Desktop.Trees.ITreeNodeCollection, depending upon whether a single-root or multiple-roots are desired.

As far as implementing a data-source, application developers may either 1) create custom helper classes that implement ITreeNode and ITreeNodeCollection and wrap an underlying object graph, or 2) use the generic TreeNode<TItem> and TreeNodeCollection<TItem> classes, which rely heavily on the use of delegates to extract information from an underlying object graph.  Use of approach 2 with anonymous delegates can result in an extremely concise mapping of an object graph to a tree, although the code tends to look like Lisp.

Also: Renamed ComparerWrapper.cs, EnumerableWrapper.cs, EnumeratorWrapper.cs, and ListWrapper.cs to TypeSafeComparerWrapper.cs, TypeSafeEnumerableWrapper.cs, TypeSafeEnumeratorWrapper.cs, and TypeSafeListWrapper.cs respectively, to more accurately reflect the purpose of these classes.