Bug: #228
Renamed IActionBuilder and ActionBuilder to IActionBuildingContext and ActionBuildingContext respectively.  Moved the logic that actually builds the actions out of the builder class and into the attribute themselves.  This makes the framework extensible by allowing new attributes to be derived from ActionAttribute.  The processor will automatically process these new attributes in the same way that it does the built-in ones.

This is part 1 of this fix - more to follow.