Requirement: R1427
Added StandardSettingsProvider class.  This class looks for an extension of EnterpriseConfigurationStoreExtensionPoint.  If it finds an extension, it calls IEnterpriseConfigurationStore.GetSettingsProvider and delegates to that settings provider.  If it does not find this extension, it delegates to LocalFileSettingsProvider.  This allows the application to be deployed in either an enterprise or standalone environment, with settings automatically persisted to the correct storage.

Any settings class that wishes to take advantage of the enterprise configuration storage must specify the StandardSettingsProvider using the attribute:

[SettingsProvider(typeof(ClearCanvas.Common.Configuration.StandardSettingsProvider))]