Requirement: R1627
Created the BackgroundTask class.  This is a lightweight wrapper around the System.ComponentModel.BackgroundWorker class, that provides some minor enhancements and a clearer API (because the BackgroundWorker API is confusing in my opinion).

The BackgroundTask provides a simple means for an ApplicationComponent or Tool to execute a task in the background without dealing with the intricacies of multi-threading.  It is meant to be used at the application level (hence inclusion in the Desktop project) and not as a general purpose multi-threading option.

Usage is as follows:

// create task
BackgroundTask task = new BackgroundTask(
   delegate(IBackgroundTaskContext context)
   {
      while(...)
      {
        // do some more background processing
        context.ReportProgress(...);
      }
      context.Complete(...);
   });

// hook up feedback events
task.ProgressUpdated += ProgressHandler;
task.Terminated += TerminatedHandler;

// start task
task.Run();