Requirement: R1901 - Implement a server-based time-provider
Added ClearCanvas.Ris.Client.Common.EnterpriseTimeProvider, an implementation of Common.Platform.TimeProviderExtensionPoint

ClearCanvas.Common.Platform.Time will use this plugin instead of the local machine time when the plugin is installed

The plugin provides a consistent time to the application from the enterprise datasource specified in the nHibernate configuration,  The initial call to CurrentTime will query the database to calculate an offset from the local machine time.  If the database query fails, the local time is returned, and subsequent calls to CurrentTime will attempt to re-query the database.  If the database query is successful, subsequent calls CurrentTime will calculate the appropriate enterprise time from the local machine time using this offset.  The offset will be periodically re-calculated from the database (by default, every 60 seconds).