Bug: #373
Added overloads to IEntityBroker.Find, IEntityBroker.FindOne and IEntityBroker.Count that accept an array of SearchCriteria objects which are OR'd together.  The individual sub-criteria of the SearchCriteria are still combined using AND, as was the case before.  The overall effect is that it is now possible to create queries that combine AND and OR operators.