Requirement: R2207
Modified Platform.GetService so that any exceptions thrown when Dispose() is called on the service are logged and discarded.  According the .NET framework book, Dispose() should never throw exceptions.  However, it seems the designers of WCF disregarded this advice (or the book is wrong).  In any case, allowing exceptions to be throw from Dispose() often obscures another more relevant exception, for example:

using(IDisposable d = new SomeDisposableObject())
{
   // do some work, some goes wrong
   throw new RelevantException();
}

In the above code, when the using statement calls d.Dispose(), any exception thrown will supercede RelevantException, which is not a desirable behaviour in most cases.