Bug: #423 Interim commit for Henry to test and see if the changes reduces the magnitude of the problem.
Wrote some additional typemaps that I had neglected to do before, for the Dispose() method of wrapper classes. Basically, when a wrapper object, like T_ASC_Association is used by calling a member method and passing in another object like T_ASC_Parameters, a member variable is assigned the reference to the T_ASC_Parameters object, so that garbage collection does not take place in the course of the P/Invoke call (c.f. HandleRef in MSDN documentation).
Unfortunately, once set, it isn't always true that they are unset.

As long as the "using (...)" statement is used, the memory for the unmanaged objects should be released as soon as possible.