Bug: #387
Modified implementation of TimeService so that it just returns DateTime.Now rather than using NHibernate to ask SQL server for the time.  This is not intended to be a permanent fix.  Ideally it should obtain the time from a time-server, or from SQL server, but if using SQL server it must not rely on the NHibernate subsystem to communicate with SQL server, because the application may for various reasons need to obtain the time before the NHibernate subsystem has been initialized.