Bug: #496 - Duplicate SOPInstance_ records are created if LocalDataStore Service processes same image more than once
- Verified after import that this occurs using rebuild via LDS and the RebuildDatabase Utility; the same files will be entered in the database again and again.
- The problem was that when a new SopInstance was added to the parent Series' ISet of SopInstances, the new sop was being compared with the existing sops in the set using the ImageSopInstance.Equals() method.  However, due to a logic error, that method always returned false and each new image was always added whether it was a duplicate or not.
- I have changed the database schema to have a not-null & unique constraint on the Study, Series and Sop Uids in each of the corresponding tables and have also fixed the Equals method of the ImageSopInstance class.

Consequently, the performance when updating the database during retrieve/import/rebuild (see #420) has improved considerably in terms of the total time take for each update (now quite fast compared to before), however the CPU will still spike quite high for that time, just as before.

I would say that #420 is improved, but not resolved.  Whether or not it is adequate for 1.0 will result from beta-testing.