Requirement: R2521
Modified drag and drop so that it executes the same code as the tools.  Introduced the concept of a drop handler (IDropHandler).  Each folder now optionally defines a DropHandlerExtensionPoint, and tools may choose to extend this point in order to handle drop operations.  Modified the CheckInTool and CancelTool so that they implement extensions to the CheckedInFolder.DropHandlerExtensionPoint and CancelledFolder.DropHandlerExtensionPoint respectively.