Requirement: R2521
Modified registration workflow tools so that they will force relevant folders to refresh immediately after the tool is executed.  This is done via hard-coded references, where each tool "knows" what folder it would have affected.  This is not an ideal solution, but it is currently the only option that will offer reasonable performance (refreshing all folders every time an operation is completed would eliminate the hard-coded knowledge, but would also be very expensive).