Bug: #386 TableView selects same item multiple times when first displayed
This was a difficult bug to fix.  It seems to be related to the interaction between the TableView, application component infrastructure, and DotNetMagic.  It is basically impossible to pinpoint the real source of the problem, and even more impossible to fix it properly.

Instead, I've implemented a hacky solution:  when the internal DataGridView selection changes, this event is handled internally by the TableView, and rather than immediately publishing the selection changed event, a timer is used to create a 50ms delay.  If the internal selection changes multiple times, the timer is simply reset.  The overall effect is to create a kind of "smoothing" function, whereby rapid selection changes (multiple changes within 50 ms) are not propagated to the client code.

Although this is a hacky solution, it seems to work, and even offers the added benefit that if the user changes the selection rapidly (ie by holding down the arrow key) the application will only respond when the selection stabilizes.  However, as with all hacky solutions, there is always the possibility that subtle bugs may be introduced, so keep an eye out for this.