Bug: #555  Classes derived from PersistentFsm throw NHibernate exceptions
This bug is caused because Castle DynamicProxy library, used by NHibernate to create proxies, does not support generics, and PersistentFsm is a generic class.  The next version of DynamicProxy is supposed to add support for generics.

In the interim, I have modified the implementation of PersistenceContext so as to minimize the creation of proxies and instantiate the real objects directly when possible.  This probably should have been done anyways, as an optimization (the previous implementation would always create a proxy, whether it one was requested or not).  However, this does not really solve the problem, just makes it less likely to occur, which is probably good enough for the time being.