Interim Commit for Bugs #353 and #587 and task #602
Changed OffisDcm_typemaps.i to add a typemap from the C datatype unsigned long to the C# datatype uint.  This allows several areas in the Offis API where unsigned long* parameters are used to be converted to "ref uint" parameters.  This will aid in using the findAndGet routines in DcmItem for retrieving arrays of binary data.

Also, added findAndGetSequentItemFromItem and findAndGetElementFromItem to OffisDcm.  The second method can be used when retrieving string attributes to get the DcmElement for a particular tag.  The getLength() method of DcmElement can then be called to get the length of the tag, so that a StringBuilder can be allocated of the correct size before retrieving a tag.  The first routine allows easy access to individual sequences within an item and is needed for ticket #602.