Requirement: #671
Modified public accessors to hosted application components (e.g. Workspace.Component, Shelf.Component) to have type object, rather than type IApplicationComponent.  Ideally, the hosted component would not be publicly exposed at all.  However, it seems there are some scenarios where it is necessary for code outside of the desktop to access a hosted component.  Exposing it as an IApplicationComponent may lead people to believe they can consume the IApplicationComponent interface, which they should not.  Therefore, I think the best compromise is to expose it as an object.  The calling code would then cast the object to the appropriate type (not IApplicationComponent).