Intermediate commit for ticket 564: "Add Sentinelle's modified splashscreen code"
The existing splash screen code has been split into the SplashScreen and SplashScreenManager classes.  SplashScreen is simply a UI (View) class, while SplashScreenManager contains the Model logic for controlling a SplashScreen object (thread instantiation, fade in, update, fade out, thread destruction).

The existing splash screen image was previously hard-coded in SplashScreen.resx.  There was a Splash.jpg file in WinForms, but it wasn't actually being referenced anywhere.  It now exists as Splash.png file in the new WinForms/Resources folder, and is directly referred to.  Note that the .png was derived from the .jpg, and so contains JPEG-esque artifacting.  It may be best to recreate Splash.png from whatever lossless (.psd?) format the splash screen was originally built in.

The splash screen doesn't appear in Debug mode, though I would argue that it should, mainly because it's my personal opinion that Debug and Release should be identical.  Plus, the splash screen doesn't really add any significant time to debug loading - the desktop form appears at the same moment regardless, and you're free to interact with it.

The splash screen is no longer a 'topmost' window.  No need for this as it generally irritates people if they've switched to another application and the splash screen occludes it.  It's simply a 'frontmost' window within ClearCanvas itself.

Note the fancy drop shadow effect.  This is accomplished through the new BitmapOverlayForm class, a UI class designed to overlay a bitmap on top of an existing window.  Notice the new SplashShadow.png resource - that's the translucent image that's effectively overlaid on top of the splash screen to create the drop shadow effect.  Tricky, tricky.

One more commit should follow this to a) clean up some commented-out code and b) maybe tease some classes apart to make the splash screen easier to override/modify.

Secret insider tip: Press Alt-F4 when the splash screen is up for a special drop shadow surprise.