Requirement: #752 ApplicationComponent API redesign
It turned out that this was easier to fix than first thought.  Rather than modifying the API, the existing API was made to behave correctly by modifying the implementation of ApplicationComponent.CanExit.  Instead of presenting the user with a "Save changes before closing?", Yes-No-Cancel message box, the message box now says "Discard changes and close?", Ok-Cancel.  e.g. there are only 2 choices instead of 3.  By removing the "save" option, the API works correctly and is in fact more robust.  Also, there is (hopefully) little impact on usability, because if the user wants to save the changes, they just cancel and then explicitly save.  Although this requires 2 steps instead of one, it is much better from an implementation standpoint, because the component may need to validate data before saving.  In summary, the previous implementation was really just trying to be too smart and do more than was really possible.

Also added a DesktopWindow.ShowMessageBox overload that accepts a title for the message box.  The application name is automatically prepended to this title (previously the title was just the application name).  Modified ApplicationComponentHost so that it automatically supplies the title of the component as the title to the message box.  This aids the user in identifying which component is displaying the message box.